+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Introduced an `Element.modify { ... }` conditional, to allow changing properties on an element.
+
 ### Removed
 
 ### Changed