+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Image` now provides an override to prevent VoiceOver from generating accessibility descriptions.
+
 ### Removed
 
 ### Changed